Refactor code to support deploymentBucket object representation globally
