+++ lib/plugins/aws/deploy/index.js
@@ -9,7 +9,6 @@
  */
 
 const BbPromise = require('bluebird');
-const setDeploymentBucketObject = require('../lib/setDeploymentBucketObject');
 const extendedValidate = require('./lib/extendedValidate');
 const monitorStack = require('../lib/monitorStack');
 const createStack = require('./lib/createStack');
@@ -32,7 +31,6 @@ class AwsDeploy {
 
     Object.assign(
       this,
-      setDeploymentBucketObject,
       extendedValidate,
       createStack,
       setBucketName,
@@ -72,7 +70,6 @@ class AwsDeploy {
     this.hooks = {
       'before:deploy:deploy': () => BbPromise.bind(this)
         .then(() => this.serverless.pluginManager.spawn('aws:common:validate'))
-        .then(this.setDeploymentBucketObject)
         .then(() => {
           if (!this.options.package && !this.serverless.service.package.path) {
             return this.extendedValidate();
