+++ lib/plugins/aws/deploy/index.test.js
@@ -113,15 +113,12 @@ describe('AwsDeploy', () => {
     });
 
     describe('"before:deploy:deploy" hook', () => {
-      let setDeploymentBucketObjectStub;
       let extendedValidateStub;
       let spawnPackageStub;
       let spawnAwsCommonValidateStub;
       let spawnAwsCommonMoveArtifactsToTemp;
 
       beforeEach(() => {
-        setDeploymentBucketObjectStub = sinon
-          .stub(awsDeploy, 'setDeploymentBucketObject').resolves();
         extendedValidateStub = sinon
           .stub(awsDeploy, 'extendedValidate').resolves();
         spawnPackageStub = spawnStub.withArgs('package').resolves();
@@ -131,7 +128,6 @@ describe('AwsDeploy', () => {
       });
 
       afterEach(() => {
-        awsDeploy.setDeploymentBucketObject.restore();
         awsDeploy.extendedValidate.restore();
       });
 
@@ -141,9 +137,7 @@ describe('AwsDeploy', () => {
 
         return awsDeploy.hooks['before:deploy:deploy']().then(() => {
           expect(spawnAwsCommonValidateStub.calledOnce).to.equal(true);
-          expect(setDeploymentBucketObjectStub.calledAfter(spawnAwsCommonValidateStub))
-            .to.equal(true);
-          expect(extendedValidateStub.calledAfter(setDeploymentBucketObjectStub)).to.equal(true);
+          expect(extendedValidateStub.calledAfter(spawnAwsCommonValidateStub)).to.equal(true);
           expect(spawnAwsCommonMoveArtifactsToTemp.calledOnce).to.equal(false);
         });
       });
@@ -154,9 +148,7 @@ describe('AwsDeploy', () => {
 
         return awsDeploy.hooks['before:deploy:deploy']().then(() => {
           expect(spawnAwsCommonValidateStub.calledOnce).to.equal(true);
-          expect(setDeploymentBucketObjectStub.calledAfter(spawnAwsCommonValidateStub))
-            .to.equal(true);
-          expect(spawnAwsCommonMoveArtifactsToTemp.calledAfter(setDeploymentBucketObjectStub))
+          expect(spawnAwsCommonMoveArtifactsToTemp.calledAfter(spawnAwsCommonValidateStub))
             .to.equal(true);
           expect(extendedValidateStub.calledAfter(spawnAwsCommonMoveArtifactsToTemp))
             .to.equal(true);
