+++ lib/plugins/aws/package/index.js
@@ -3,7 +3,6 @@
 const BbPromise = require('bluebird');
 const path = require('path');
 const mergeCustomProviderResources = require('./lib/mergeCustomProviderResources');
-const setDeploymentBucketObject = require('../lib/setDeploymentBucketObject');
 const generateArtifactDirectoryName = require('./lib/generateArtifactDirectoryName');
 const generateCoreTemplate = require('./lib/generateCoreTemplate');
 const saveServiceState = require('./lib/saveServiceState');
@@ -22,7 +21,6 @@ class AwsPackage {
 
     Object.assign(
       this,
-      setDeploymentBucketObject,
       generateCoreTemplate,
       mergeIamTemplates,
       generateArtifactDirectoryName,
@@ -59,7 +57,6 @@ class AwsPackage {
         .then(() => this.serverless.pluginManager.spawn('aws:common:cleanupTempDir')),
 
       'package:initialize': () => BbPromise.bind(this)
-        .then(this.setDeploymentBucketObject)
         .then(this.generateCoreTemplate),
 
       'package:setupProviderConfiguration': () => BbPromise.bind(this)
