+++ lib/plugins/aws/package/index.test.js
@@ -77,7 +77,6 @@ describe('AwsPackage', () => {
 
   describe('hooks', () => {
     let spawnStub;
-    let setDeploymentBucketObjectStub;
     let generateCoreTemplateStub;
     let mergeIamTemplatesStub;
     let generateArtifactDirectoryNameStub;
@@ -88,8 +87,6 @@ describe('AwsPackage', () => {
     beforeEach(() => {
       spawnStub = sinon
         .stub(serverless.pluginManager, 'spawn');
-      setDeploymentBucketObjectStub = sinon
-        .stub(awsPackage, 'setDeploymentBucketObject').resolves();
       generateCoreTemplateStub = sinon
         .stub(awsPackage, 'generateCoreTemplate').resolves();
       mergeIamTemplatesStub = sinon
@@ -106,7 +103,6 @@ describe('AwsPackage', () => {
 
     afterEach(() => {
       serverless.pluginManager.spawn.restore();
-      awsPackage.setDeploymentBucketObject.restore();
       awsPackage.generateCoreTemplate.restore();
       awsPackage.mergeIamTemplates.restore();
       awsPackage.generateArtifactDirectoryName.restore();
@@ -129,9 +125,7 @@ describe('AwsPackage', () => {
 
     it('should run "package:initialize" hook', () => awsPackage
       .hooks['package:initialize']().then(() => {
-        expect(setDeploymentBucketObjectStub.calledOnce).to.equal(true);
-        expect(generateCoreTemplateStub.calledAfter(setDeploymentBucketObjectStub))
-          .to.equal(true);
+        expect(generateCoreTemplateStub.calledOnce).to.equal(true);
       })
     );
