+++ lib/plugins/aws/provider/awsProvider.js
@@ -120,6 +120,21 @@ class AwsProvider {
     if (timeout) {
       AWS.config.httpOptions.timeout = parseInt(timeout, 10);
     }
+
+    // Support deploymentBucket configuration as an object
+    const provider = this.serverless.service.provider;
+    if (provider && provider.deploymentBucket) {
+      if (_.isObject(provider.deploymentBucket)) {
+        // store the object in a new variable so that it can be reused later on
+        provider.deploymentBucketObject = provider.deploymentBucket;
+        if (provider.deploymentBucket.name) {
+          // (re)set the value of the deploymentBucket property to the name (which is a string)
+          provider.deploymentBucket = provider.deploymentBucket.name;
+        } else {
+          provider.deploymentBucket = null;
+        }
+      }
+    }
   }
 
   request(service, method, params) {
